+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.6.0 - 2016-02-18
+
+### Added
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#86](https://github.com/zendframework/zend-inputfilter/pull/86),
+  [#95](https://github.com/zendframework/zend-inputfilter/pull/95), and
+  [#96](https://github.com/zendframework/zend-inputfilter/pull/96) update the
+  component to be forwards-compatible with zend-servicemanager v3.
+
 ## 2.5.6 - TBD
 
 ### Added